+++ .prettierrc
@@ -1,6 +1,5 @@
 {
   "endOfLine": "lf",
-  "insertFinalNewline": true,
   "printWidth": 100,
   "trailingComma": "all",
   "tabWidth": 2,
@@ -10,5 +9,20 @@
   "bracketSameLine": true,
   "arrowParens": "always",
   "useTabs": false,
-  "plugins": ["prettier-plugin-tailwindcss"]
+  "importOrder": [
+    "^react$",
+    "^next(.*)$",
+    "<THIRD_PARTY_MODULES>",
+    "^@/app/(.*)$",
+    "^@/components/(.*)$",
+    "^@/layouts/(.*)$",
+    "^@/hooks/(.*)$",
+    "^@/lib/(.*)$",
+    "^@/styles/(.*)$",
+    "^@/(.*)$",
+    "^[./]"
+  ],
+  "importOrderSeparation": true,
+  "importOrderSortSpecifiers": true,
+  "plugins": ["@trivago/prettier-plugin-sort-imports", "prettier-plugin-tailwindcss"]
 }
