+++ src/app/_components/home/ImageBox.tsx
@@ -1,5 +1,6 @@
-import Image from 'next/image';
 import React from 'react';
+
+import Image from 'next/image';
 
 interface Props {
   title: string;
