+++ src/app/_components/home/Introduce.tsx
@@ -1,4 +1,5 @@
 import Image from 'next/image';
+
 import ArrowDown from '../../../../public/assets/icons/ArrowUpBlue.svg';
 import WithCotamButton from './WithCotamButton';
