+++ src/app/recruit/_components/ContactSection.tsx
@@ -1,4 +1,5 @@
 import Button from '@/components/common/Button';
+
 import { KAKAO_OPEN_CHAT_URL } from '@/constants/externalLinks';
 
 const ContactSection = () => {
