+++ src/app/robots.ts
@@ -1,5 +1,6 @@
-import { SITE_URL } from '@/constants/metadata';
 import { MetadataRoute } from 'next';
+
+import { SITE_URL } from '@/constants/metadata';
 
 export default function robots(): MetadataRoute.Robots {
   return {
