+++ src/app/sitemap.ts
@@ -1,4 +1,5 @@
 import { MetadataRoute } from 'next';
+
 import { SITE_URL } from '@/constants/metadata';
 
 export default function sitemap(): MetadataRoute.Sitemap {
