+++ src/components/common/FAQAccordion.tsx
@@ -1,7 +1,8 @@
 'use client';
 
-import { cn } from '@/lib/utils';
 import { useReducer } from 'react';
+
+import { cn } from '@/lib/utils';
 
 interface Props {
   className?: string;
